+++ bootx64/Cargo.toml
@@ -23,3 +23,4 @@ x86_64 = "0.14.4"
 elf_rs = "0.1.3"
 os_units = "0.4.2"
 predefined_mmap = { path = "../libs/predefined_mmap" }
+boot_info = { path = "../libs/boot_info" }
