+++ kernel/Cargo.toml
@@ -44,6 +44,7 @@ xmas-elf = "0.8.0"
 uart_16550 = "0.2.15"
 spinning_top = { version = "0.2.4", features = ["nightly"] }
 predefined_mmap = { path = "../libs/predefined_mmap" }
+boot_info = { path = "../libs/boot_info" }
 
 [build-dependencies]
 cc = "1.0.69"
