+++ libs/boot_info/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "boot_info"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+os_units = "0.4.2"
+predefined_mmap = { path = "../predefined_mmap" }
+uefi = "0.11.0"
+vek = { version = "0.15.1", features = ["libm"], default-features = false }
+x86_64 = { version = "0.14.6", default-features = false }
