+++ libs/common/src/lib.rs
@@ -5,6 +5,4 @@
 #![deny(clippy::all)]
 
 pub mod debug;
-pub mod kernelboot;
 pub mod mem;
-pub mod vram;
